Split /go and /golink modules
