+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
 pub mod codefmt;
 pub mod go;
+pub mod golink;
 pub mod ping;
